Rename and move PrincipalConstants.
