+++ src/main/kotlin/org/wfanet/measurement/common/identity/AuthorityKeyServerInterceptor.kt
@@ -25,7 +25,7 @@ import io.grpc.ServerInterceptor
 import io.grpc.Status
 import java.security.cert.X509Certificate
 import javax.net.ssl.SSLSession
-import org.wfanet.measurement.api.PrincipalConstants
+import org.wfanet.measurement.api.v2alpha.ContextKeys
 import org.wfanet.measurement.common.crypto.authorityKeyIdentifier
 import org.wfanet.measurement.common.grpc.failGrpc
 
@@ -50,7 +50,7 @@ class AuthorityKeyServerInterceptor : ServerInterceptor {
     headers: Metadata,
     next: ServerCallHandler<ReqT, RespT>
   ): ServerCall.Listener<ReqT> {
-    if (PrincipalConstants.PRINCIPAL_CONTEXT_KEY.get() != null) {
+    if (ContextKeys.PRINCIPAL_CONTEXT_KEY.get() != null) {
       return Contexts.interceptCall(Context.current(), call, headers, next)
     }
