+++ src/main/kotlin/org/wfanet/measurement/common/identity/BUILD.bazel
@@ -13,7 +13,7 @@ kt_jvm_library(
         "DuchyInfo.kt",
     ],
     deps = [
-        "//src/main/kotlin/org/wfanet/measurement/api:principal_constants",
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:context_keys",
         "//src/main/proto/wfa/measurement/config:duchy_cert_config_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/io/grpc:api",
         "@wfa_common_jvm//imports/java/io/grpc/stub",
